moved Header.js into components folder and updated import of Header.js and logo
