+++ starwars/src/App.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import axios from "axios";
 import "./App.css";
 import CharacterCard from "./components/Character";
-import Header from "./Header"
+import Header from "./components/Header"
 const App = () => {
   // Try to think through what state you'll need for this app before starting. Then build out
   // the state properties here.
